// apps/recipes-client-trpc/src/app/components/interfaces/recipes.ts
export interface Recipes {
    recipeId: string;
    title: string;
    category: string;
    sensitivity: string;
    creatorName: string;
    creatorId: string;
    creatorImage: string;
    rating: number;
    countyOfOrigin: string;
    difficulty: string;
    image: string;
    ingredients: string[];
    instructions: string;
    preparationTime: string;
    createdAt: string;
}